Added more tasks to to-do list
Specifically expanding on the task of restricting the frog's positions. A drawback to using hexagons instead of squares: if squares were used then it would've been pretty straightforward to move left or right. With hexagons, since they're not painted side to side, but rather in alternating positions, if the user wants to move left or right then they'd be moved to the top-left/right hexagon.